+++ realtime/src/main/java/cn/leancloud/im/v2/LCIMTypedMessage.java
@@ -236,21 +236,6 @@ public class LCIMTypedMessage extends LCIMMessage {
    * 从 conversation 数据中解析 lastMessage
    * @return
    */
-//  private static AVIMMessage parseMessage(String conversationId, JSONObject jsonObject) {
-//    if (null != jsonObject && jsonObject.containsKey(KEY_MESSAGE_ID)) {
-//      try {
-//        String from = jsonObject.getString(KEY_MESSAGE_FROM);
-//        String data = jsonObject.getString(KEY_MESSAGE_CONTENT);
-//        long timestamp = jsonObject.getLong(KEY_MESSAGE_TIMESTAMP);
-//        String msgId = jsonObject.getString(KEY_MESSAGE_ID);
-//        AVIMMessage message = new AVIMMessage(conversationId, from, timestamp, -1);
-//        message.setMessageId(msgId);
-//        message.setContent(data);
-//        return AVIMMessageManagerHelper.parseTypedMessage(message);
-//      } catch (Exception e) {}
-//    }
-//    return null;
-//  }
   static LCIMMessage parseMessage(String conversationId, Map<String, Object> jsonObject) {
     if (null != jsonObject && jsonObject.containsKey(KEY_MESSAGE_ID)) {
       try {